refactor: move scss files out of css folder
