+++ src/components/ComposerSessionIndicator.vue
@@ -97,7 +97,7 @@ export default {
 </script>
 
 <style lang="scss" scoped>
-@use '../../css/variables.scss';
+@use '../styles/variables.scss';
 
 .composer-session {
 	position: fixed;
