+++ src/components/Message.vue
@@ -198,7 +198,7 @@ export default {
 </script>
 
 <style lang="scss" scoped>
-@use '../../css/variables.scss';
+@use '../styles/variables.scss';
 
 .v-popover > .trigger > .action-item {
 	border-radius: 22px;
