+++ src/components/MessageAttachment.vue
@@ -278,7 +278,7 @@ export default {
 </script>
 
 <style lang="scss" scoped>
-@use '../../css/variables.scss';
+@use '../styles/variables.scss';
 
 @media screen and (max-width: #{variables.$breakpoint-mobile}) {
 	.attachment{
