+++ src/styles/mail.scss
